Fix UnhandledRejection error when failed to start application on iOS during debug
In case we are unable to start application on iOS device just before attaching a debugger, CLI will raise unhandled rejection.
The problem is that we want to start two actions in parallel - start of application and attaching a debugger.
We do not await the start application and when an error is raised in it, Unhandled Rejection is raised.
In order to prevent this, use Promise.all - this way in case any of the operations fail, we'll receive it as error in the same method.