Fix application is not started on iOS Simulator every other time
In case you are using iOS Simulator with version < 10, every second LiveSync operation does not start the app on device. The problem is that calling `simctl terminate` is not killing the app immediately.
However the action is executed, so we call start application. After its been called, the terminate succeeds and kills the app. So the app is not running. When a change is applied we detect app is not running and just start it.
So it works. Next change calls terminate and again the app seems like it has not been started.

The fix is in ios-sim-portable, so update its version.