Fix tracking of Node.js version, LiveSync action and user sessions count (#3117)
Fix tracking of LiveSync action - it has been always tracked even when Build is required, so the workflow of user seemed like:
- Build
- LiveSync
- Deploy
- LiveSync

Now it will be:
- Build
- Deploy
- LiveSync

Also update common lib, where the following changes are applied:

Fix tracking of Node.js version in Eqatec and user sessions count

Whenever a new eqatec monitor is started, we should track the version of Node.js. However, the current code was calling `trackFeatureCore` which always tracks to the Eqatec Analytics projects used for Feature trackings.
This way, when any eqatec monitor is started, we've been sending information for Node.js version to the feature tracking projects. Fix this by passing the api key of the project in which we want to track the information.

Also fix the user sessions count - whenever `start` method is called, we are constructing analytics settings (which increases the counter of current user sessions with 1) and pass this to `startEqatecMonitor` method.
At this point we check if we've already started this monitor and in case not, we do not call start again. However, the values in `user-settings.json` for sessions count remains incorrect. Fix this by adding the check for already started monitor in the beginning of `start` method.

Add unit tests to verify tracking in multiple projects. Change some of the already exising verifications to use arrays instead of strings for easier comparison.