Do not use spinner in CI environment
In case terminal is not interactive, the spinner we are using during `tns doctor` or `tns platform add ...` is printing a lot of messages and polutes the output.
So in case we are in non-interactive environment, print the spinner message a single time. In order to achieve this, use the newly introduce method in $progressIndicatior from submodule;

Add getSpinner method to $progressIndicator

Add `getSpinner` method to $progressIndicator - it will return a new instance of clui.Spinner in case terminal is interactive. In case it is not - a mocked instance will be returned.
This way in CI builds the spinner will not print tons of repeated messages.