fix(debug): debug-brk option does not work for iOS Simulator (#3340)
The `debug-brk` option was broken in a previous commit as we were trying to connect to the backend port (18181). The problem is that connecting to it makes the runtime thinks debugger is attached and it continues to next statements.
Instead of connecting to the port, check if it is in LISTEN state.