feat(create): Keep pluginsData key from template's package.json
In case project template has a preconfigured plugins data (for example API keys), it will be stored in the template's package.json.
After creating the project, CLI removes all keys except the one defined in `PackageJsonKeysToKeep` array from `app/package.json` file.
So it removes the predefined configuration of the pluginData.
Fix this by adding the entry "pluginsData" in the array, so we'll keep it as well.