feat(nativescript): Allow calling Kinvey.init without arguments
We are trying to create a convention for all NativeScript plugins to keep configuration data in `<project dir>/app/package.json`.
The values of the configuration data can be safely retrieved at runtime.
The idea of this PR is to read appKey, appSecret and masterSecret from this package.json file. The format there will be:
```JSON
{
	"android": {
		"v8Flags": "--expose_gc"
	},
	"main": "main.js",
	"name": "tns-template-hello-world-ng",
	"version": "3.1.1",
	"pluginsData": {
		"kinvey-nativescript-sdk": {
			"config": {
				"appKey": "app_key_value",
				"appSecret": "app_key_secret"
			}
		}
	}
}
```

This improves the logic on clients side - the init method can be called without arguments.
The other important improvement is that in Sidekick we can check if there are values for appKey and appSecret in this well known place and point the user to enter them in case they are missing. Also this allows us to easily change their values, i.e. support edit of values.

The current code change will allow the users to pass config to `Kinvey.init` method, so it is fully backwards compatible. However, in case some values are missing, we'll try to get them from package.json.