refactor: Set logFilter's project name when starting iOS application
When strating iOS Application, set the projectName to the log filter, so
we can filter the logs by it. Same logic is used for Android, where we
are using the PID of the started application. So introduce
IDeviceLogOptions, where logLevel, pid and projectName can be set. Pass
the projectName wherever is required in order to get it in
startApplication.