fix: Plugin remove command fails in case plugin name has dot
In case plugin name has dot (`.`) in its name, the `plugin remove` command fails.
The reason is the way we are trying to remove plugin related data from the `nativescript` key of application's package.json.
We are using `.` to concatenate the keys inside nativescript key and later we are trying to split by `.` in order to find which key should be deleted.
This works in most of the cases, but when the plugin name has `.` in it, our logic fails as the split operation finds the dot.
Fix this by using some symbols that are not allowed in npm for plugins names. Use symbols that will rarely be used in the same sequence.