fix(logs): Logs from iOS device are not shown on Windows
When cloud build is used and iOS device does not have Developer Disk Image mounted, the logs are not visible. The problem is that the code fails to start/stop the application before starting the device log streams.
Fix this by starting the logs before trying to start/stop the application.