fix: Getting started prompter fails
The prompter for getting started changes fails in some cases, as we do not await the validation of the livesyncCommandHelper.validate call. So we try to execute some operations as we think the environment is setup correctly, but we are unable to do so and the code fails.

Also add additional option to prompter in case cloud extension is installed. This gives better visibility of the feature