fix(analytics): Message for tracking in GA is always printed
In case the usage reporting is disabled, the verbose log of CLI still prints that will send information to Google Analytics. In fact it does not do it. The problem is that the message is printed before checking the state of the usage reporting.
Fix this by placing the printing in the correct method.