fix: Support Node.js 10.x.x
Add support for Node.js 10.x.x. Currently, when trying to use CLI with Node.js 10.x.x, error is thrown. The problem is that we are using npm-shrinkwrap.json to keep using same versions of all dependencies, no matter how they are defined in the package.json files.
The current problem is that the `unzip` package relies on `fstream` package, which depends on `graceful-fs` 3.0.11 version. This version of `graceful-fs` depends on `natives` package by using the range `^1.1.0`. In our `npm-shrinkwrap.json` this range is resolved to version 1.1.2. This version of the `natives` package does not work with Node.js 10.x.x. There's a newer one - 1.1.3 that fixes it. So the solution is to update the npm-shrinkwrap.json and use the latest version of `natives` package.

Also add version 10.x.x as supported in CLI's package.json, so we'll not show warning that it is not validated.