fix: Build plugins when path to project has space
In case the path to project has space in the name and there's Android plugin that should be build, the build operation fails. The reason is usage of exec - change it to spawn, which handles the spaces correctly.

feat: Add `buildAndroidPlugin` hook, which is executed just before spawning the gradle build for specific plugin. The hook can be used to modify plugin's structure before building the plugin.