feat: Deprecate support for Node.js 6.x.x
As Node.js 6.x.x is in maintenance mode, we deprecate support for it. We'll keep it deprecated until the usage of it is very low.

Disable support for Node.js 4.x.x and 5.x.x

Currently, when CLI is used with Node.js 4.x.x or 5.x.x, a warning has been shown. As we are deprecating support for Node.js 6.x.x and Node.js 4.x.x is near the end of its life, remove the support for Node.js versions below 6.x.x
In case Node.js 4.x.x is used, CLI will show error and will not execute the command. Node.js 4.x.x has been deprecated for some months.