feat: Deprecate support for all macOS versions below High Sierra
Add deprecation message in case the current macOS version is below High Sierra. As the latest Xcode versions can be installed only on this OS and we state that we support only latest Xcode, we have to deprecate the support for other OS versions.
The deprecation message will be shown on each CLI command.
As the deprecation messages from CLI should be shown in Sidekick as well, expose new methods in CLI's `sysInfo` moodule, so Sidekick can get the warnings and show them.
In order to get the warnings for Node.js version, change the way `verifyNodeVersion` is called.
Expose method for getting the currently supported Node.js version range.
Add tests for the new methods.