fix: Detect iOS devices connected by USB only
Currently CLI detects iOS Devices with Wi-Fi sync enabled. However, we can only detect them, but we are not able to deploy and livesync applications on them. This also breaks the workflow with devices connected by USB, but with Wi-Fi sync enabled.
So detect only the USB connected devices and disregard the ones connected with Wi-Fi sync enabled. This is handled in the 0.4.11 version of ios-device-lib, so just update the version