fix: Unable to fast sync changes with iOS runtime 4.0.1 and earlier
CLI should be able to execute fast sync on devices and iOS Simulator (single instance) when using older runtime version (4.0.1 and prior).
However, currently it just hangs as the promise that should return the port for fast sync, is never resolved. Fix this by resolving the promise in this case.