fix: Unable to debug on iOS Simulator when multiple sims are running
In case multiple iOS Simulators are running and `tns debug ios` is called, CLI prompts you to select you on which device/simulator to debug.
Selecting the device and trying to debug your application leads to error as CLI does not pass correctly the device identifier.
Fix this by passing the correct device identifier.
Also update ios-sim-portable to latest version, where several fixes for this case are applied