fix: `tns debug ios --start` fails
The `tns debug ios --start` command fails as we do not get the port from iOS Application on device in this case. In order to fix this, start the log parsing in case --start option is passed.
Also fix the attaching to ATTACH_REQUEST_EVENT_NAME - we've incorrectly attached to device's application manager, which does not emit such event. Fix it by attaching to `iOSNotification`, which is the correct one.
Also remove incorrect cache decorator and ensure we do not start logging for iOS devices more than once.