fix: platform remove command fails with old Android runtime
After upgrading `nativescript-doctor`, the version of the Java is
validated against the current Android runtime version. In case the
runtime version cannot work with Java 10, `nativescript-doctor` does not
allow the operation to continue. This breaks the `platform remove`
command, as it checks the system requirements first, before executing
the actual removal. However, this is no longer required as the code for
removing the platform is `try-catched` and in case the current
environment is not setup correctly, it will try to remove the
platforms/<platform> dir and the respective key in package.json.
So remove the environment validation from this command. Also remove the
`validatePlatformInstalled` call from the `canExecute` method of the
command, as it is actually called in the `removePlatforms` method.