fix: Validate environment based on specified runtime version
Add logic to validate environment based on the specified runtime version. This is required for the cases where the runtime version is not specified in the project's package.json but we need to verify if a runtime version can be used
For `tns platform update <platform>` command validate the runtime version that will be installed based on the passed argument. This has two cases:
1. In case version is specified: `tns platform update android@4.0.1` - we'll get the 4.0.1 version and validate if it can be used with currently installed Java compiler version.
2. In case version is not specified: `tns platform update android` - we'll not validate the java version here as we know we'll install at least 4.1.0 runtime, which is compatible with all Java versions

For `tns platform clean <platform>` command we need to validate the current platform immediately as the command removes it and adds the same version after that.