fix: Device logs are not properly shown with debug command
The console.log from application should be shown in the terminal when `tns debug <platform>` command is executed.
When `tns debug android --start` is used, the console.log messages are not visible as noone has started reading them. Add logic to read and filter them based on the PID of the application.
When `tns debug ios [--start]` is used with older runtime (4.0.1 for example) with iOS Simulator, the device logs are not shown as the filtering is based on the projectName, but noone has set this projectName to the filter. So we filter everything. Fix this by setting the projectName to the filter instance.
Fix the `tns debug ios --justlaunch` and `tns debug ios --start --justlaunch` which are always printing console.logs, while in fact they shouldn't.