fix: Run and debug fail for iOS in Sidekick
When CLI is used as a library (in NativeScript Sidekick for example), calling methods for run and debug currently fails. The problem is with the parsing of the iOS logs, which was introduced several days ago.
The logic uses `$projectData` from the `$injector`. However, when CLI is used as a library, the `$projectData` is never initialized. Instead of using `$projectData` pass the project directory and construct new project data (from `$projectDataService`) where required.