fix: `tns debug ios` fails to show chrome url
In case when you do not have iOS Simulators running, the detection that checks if there are any simulators, takes 15 seconds. CLI executes this detection on various places.
This leads to several issues, but the main one is the `tns debug ios` command when there's no running Simulators. In this case CLI often fails to display the URL for debugging as the timeout while waiting to get the port from iOS Device's logs passes while we are trying to find running simulators.
Fix this by updating ios-sim-portable, where the detection does no longer wait 15 seconds - we just try twice for 500 ms to detect simulators and that's it.