fix: Trying to run unit test fails with latest karma
When tests are initialized, the `tns test <platform>` command fails in case karma 2.0.3 is used. In order to fix it, until karma resolves the issue on their side, hardcode the version of karma that is installed by CLI on `tns test init` command.

Also install mocha when it is selected for testing framework. In case we do not install it, the test command fails again.