fix(livesync): stop forcing native prepare when using --bundle
These checks are no longer needed. They also cause the CLI to
rebuild the .aar files on every change when using `tns run android
--bundle`.

- The first check:
The CLI watches the `App_Resources` folder and will execute the logic
for preparing them whenever there's a change.

- The second check:
The CLI will try to detect changes in node_modules by comparing the
timestamps of every file there on every change. When there's a change,
it will set the `modulesChanged` flag to true. Therefore, the `bundle` flag is not needed.

fixes #3707